+++ ibvn/Modules/ElRetoDeHoy/Model/YouTubeList.swift
@@ -8,43 +8,43 @@
 import Foundation
 
 // MARK: - YoutubeSearch
-struct YouTubeList: Codable, Identifiable {
-    let id = UUID()
-    let kind, etag, nextPageToken: String
-    let pageInfo: PageInfo
-    let items: [ListItem]
-    
-    init(kind: String = "",
-         etag: String = "",
-         nextPageToken: String = "",
-         pageInfo: PageInfo = .init(),
-         items: [ListItem] = []) {
-        self.kind = kind
-        self.etag = etag
-        self.nextPageToken = nextPageToken
-        self.pageInfo = pageInfo
-        self.items = items
-    }
-    
-    enum CodingKeys: CodingKey {
-        case kind
-        case etag
-        case nextPageToken
-        case pageInfo
-        case items
-    }
-}
-
-// MARK: - Item
-struct ListItem: Codable, Identifiable {
-    let kind: String
-    let etag, id: String
-    let snippet: Snippet
-    
-    enum CodingKeys: String, CodingKey {
-        case kind
-        case etag
-        case id
-        case snippet
-    }
-}
+//struct YouTubeList: Codable, Identifiable {
+//    let id = UUID()
+//    let kind, etag, nextPageToken: String
+//    let pageInfo: PageInfo
+//    let items: [ListItem]
+//    
+//    init(kind: String = "",
+//         etag: String = "",
+//         nextPageToken: String = "",
+//         pageInfo: PageInfo = .init(),
+//         items: [ListItem] = []) {
+//        self.kind = kind
+//        self.etag = etag
+//        self.nextPageToken = nextPageToken
+//        self.pageInfo = pageInfo
+//        self.items = items
+//    }
+//    
+//    enum CodingKeys: CodingKey {
+//        case kind
+//        case etag
+//        case nextPageToken
+//        case pageInfo
+//        case items
+//    }
+//}
+//
+//// MARK: - Item
+//struct ListItem: Codable, Identifiable {
+//    let kind: String
+//    let etag, id: String
+//    let snippet: Snippet
+//    
+//    enum CodingKeys: String, CodingKey {
+//        case kind
+//        case etag
+//        case id
+//        case snippet
+//    }
+//}
